fix migrations and add uploaded images

+++ database/migrations/2026_05_08_105317_add_posisi_foto_to_gurus_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('gurus', function (Blueprint $table) {
-            $table->string('posisi_foto')->default('50% 20%')->after('foto');
+            if (!Schema::hasColumn('gurus', 'posisi_foto')) {
+                $table->string('posisi_foto')->default('50% 20%')->after('foto');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('gurus', function (Blueprint $table) {
-            $table->dropColumn('posisi_foto');
+            if (Schema::hasColumn('gurus', 'posisi_foto')) {
+                $table->dropColumn('posisi_foto');
+            }
         });
     }
 };
